requirements.txt
flask==3.0.2
werkzeug==3.0.3
hypercorn==0.16.0
aiomysql==0.2.0
python-dotenv==1.0.0
asgiref==3.7.2
cryptography==42.0.5
watchfiles==0.18.1
pyjwt==2.8.0
bcrypt~=4.3.0
